requirements_rpi.txt
# PyTorch and related packages (install via pip with CPU-only builds)
torch
torchvision
torchaudio

# LWCC library for crowd counting
lwcc

# Computer vision and image processing
opencv-python
Pillow

# Math and scientific computing
numpy
scipy

# System monitoring and performance
psutil

# Camera and GPIO (if needed)
picamera2

# Display and GUI (optional)
matplotlib

# Development and debugging
tqdm 